The sidebar component has been adapted to the next view of the application: HrStudentProfileView.
